Set nunjucks noCache within app.js instead (#24)
Setting `nunjucksEnv.opts.noCache = true` after the Nunjucks environment
has been initialised doesn't actually work.

This could be resolvable by re-initialising it, but that's more
complicated as it means recreating the list of search paths. Instead it
feels simpler to revert to setting it in `app.js` for now.

The other 2 settings are default anyway, so can be dropped.

Reverts #22