Add build script to kit (using esbuild) (#96)
This adds the compiling of SASS and javascript to the kit, using
`esbuild`.

It's an optional extra, enabled by adding an `buildOptions` config to
the `init` function with references to the files you’d like to build:

For example, to compile SASS only:

```js
NHSPrototypeKit.init({
  buildOptions: {
    entryPoints: ['assets/css/*.scss']
  },
  ...
})
```

or for both SASS and javascript:

```js
NHSPrototypeKit.init({
  buildOptions: {
    entryPoints: ['assets/css/*.scss', 'assets/javascript/*.js']
  },
  ...
})
```

It could also be run standalone by using this async function:

```js
await NHSPrototypeKit.build(true, {entryPoints: ['assets/css/*.scss']})
```

This is a replacement for `gulpfile.js` in the existing NHS prototype
kit.