Add redirect-post-to-get middleware (#107)
I forgot to add this before, as currently in the kit it’s not included
as middleware in `lib/` but instead just as a regex route within
`app.js`:
https://github.com/nhsuk/nhsuk-prototype-kit/blob/main/app.js#L229-L237

The kit works fine without this, as the auto-routes are still rendering
the matching templates on the POST requests. But you do get the 'Are you
sure you want to send a form again?' message when refreshing the page,
and seeing as the NHS prototype kit (and the GOVUK one) has always had
this, I think it’s worth adding.