Add layout for admin views
Currently the password page uses the main layout file from the prototype - which may contain navigation and other links.

This adds a very minimal admin layout which contains no links in the header or footer.

It’s also used by the 500 page, to avoid any syntax issues with the prototype layout file causing the 500 page not to render.