Update auto-routing to also work with path.njk and folder/index.njk views (#127)
The prototype kit uses view templates that are a mixture of HTML and
Nunjucks. As such, you might choose to name these files `.njk`, eg to
get improved syntax highlighting or code autocompleting.

This change updates the auto-routing middleware to work with both types
of file extension.

If there is a clash, the `.html` file will take precedence over `.njk`

Fixes #126